// src/app/shared/interfaces/auxilium/patient-center/patient-charges-by-id.interface.ts
export interface PatientChargesById {
    id: number;
    sourceid: number;
    sourcetable: string;
    patientid: number;
    itemid: number;
    invoiceno: string;
    patalpha: string;
    lastweekbilled: number;
    lastdatebilled: string;
    previousdatebilled: string;
    svcdate: string;
    stopdate: string;
    todate: string;
    isoverride: boolean;
    paponumber: string;
    assetid: number;
    diag24e: string;
    patienticdid1: number;
    patienticdid2: number;
    patienticdid3: number;
    patienticdid4: number;
    submitted: number;
    allowed: number;
    dueprimary: number;
    duesecondary: number;
    prepaid: number;
    prepaid2: number;
    pos: string;
    tos: string;
    billtype: string;
    stopcode: string;
    priceid: number;
    quantity: number;
    authdate: string;
    cmnexpire: string;
    nextbill: string;
    hcpcscode: string;
    patientpayorid1: number;
    patientpayorid2: number;
    payorid1: number;
    payorid2: number;
    salesid: number;
    referid: number;
    lastseen: string;
    mod1: string;
    mod2: string;
    mod3: string;
    mod4: string;
    physician: string;
    patientdoctorid1: number;
    physeq: string;
    cmnreq: string;
    firstcmn: string;
    nextcmn: string;
    cmnfreq: number;
    cmnformid: number;
    description: string;
    monthsrented: number;
    entrydate: string;
    operatorid: string;
    assignment: boolean;
    ytdbilled: number;
    pickupreason: string;
    narrative: string;
    outsidelab: string;
    facilityid: number;
    charges: number;
    shipstatus: string;
    shipdate: string;
    billcycle: string;
    admit: string;
    discharge: string;
    lotnumber: string;
    branchid: number;
    confirm: string;
    confirmid: string;
    orderedby: string;
    ordphone: string;
    checkno: string;
    checkno2: string;
    dateprinted: string;
    groupid: string;
    datepaid: string;
    adddate: string;
    adduserid: string;
    calories: number;
    tenmonth: string;
    otherhold: string;
    primbill: string;
    secbill: string;
    deliverytype: string;
    followup: string;
    chcup: string;
    lineno: number;
    paymenttype: string;
    paymenttype2: string;
    authcount: number;
    authused: number;
    invtcat2update: string;
    notes: string;
    inventorylocationid: number;
    inventorybinid: number;
    diagnosis: string;
    diagnosis2: string;
    diagnosis3: string;
    diagnosis4: string;
    expdate1: string;
    expdate2: string;
    code1: string;
    code2: string;
    patienticdid5: number;
    patienticdid6: number;
    patienticdid7: number;
    patienticdid8: number;
    description2: string;
    pwktypecode: string;
    pwktransmissioncode: string;
    pwkattachmentcn: string;
    pwkccn: string;
    trackingno: string;
    shipvia: string;
    quantityordered: number;
    quantityshipped: number;
    usagelimitquantity: number;
    lastuser: string;
    datedelivered: string;
    lot: boolean;
    upnqualifier: string;
    upn: string;
    level: string;
    qtyorderedbyphy: number;
    patientdoctorid2: number;
    isrollup: boolean;
    salesman1pat: string;
    salesman2pat: string;
    authreq: string;
    deliveryprint: boolean;
    inventoryvendorid: number;
    accessionno: string;
    notesexpiredate: string;
    chargeimportfile: string;
    epsdt: string;
    doctorid1: number;
}